temp_backup/temp_backup/requirements.txt
Flask>=2.2.5,<3.0
Flask-Login>=0.6.2,<0.7
boto3>=1.28.2,<2.0
opencv-python>=4.8.0.74,<5.0
Pillow>=10.0.0,<11.0
numpy>=1.24.3,<2.0
firebase-admin>=6.0.1,<7.0
openpyxl>=3.1.2,<4.0
requests>=2.31.0,<3.0
google-generativeai
Werkzeug>=3.0,<4.0
Flask-WTF>=1.1.1,<2.0